requirements.txt
streamlit
networkx
pyvis
requests
python-dotenv